refactor and combine the external accounts plugin
