--- client/modules/accounts/components/index.js
@@ -1,8 +1,0 @@
-export { default as ForgotPassword } from "./forgotPassword";
-export { default as Login } from "./login";
-export { default as LoginButtons } from "./loginButtons";
-export { default as LoginFormMessages } from "./loginFormMessages";
-export { default as MainDropdown } from "./mainDropdown";
-export { default as SignIn } from "./signIn";
-export { default as SignUp } from "./signUp";
-export { default as UpdatePasswordOverlay } from "./updatePasswordOverlay";
