--- client/modules/accounts/containers/index.js
@@ -1,5 +1,0 @@
-export { default as AuthContainer } from "./auth";
-export { default as ForgotPassword } from "./forgotPassword";
-export { default as MainDropdown } from "./mainDropdown";
-export { default as MessagesContainer } from "./messages";
-export { default as UpdatePasswordOverlay } from "./passwordOverlay";
