+++ imports/plugins/core/accounts/client/components/forgotPassword.js
