+++ imports/plugins/core/accounts/client/components/mainDropdown.js
