+++ imports/plugins/core/accounts/client/components/updatePasswordOverlay.js
