+++ imports/plugins/core/accounts/client/index.js
@@ -1,13 +1,56 @@
 export { default as AccountsDashboard } from "./components/accountsDashboard";
 export { default as AdminInviteForm } from "./components/adminInviteForm";
 export { default as EditGroup } from "./components/editGroup";
+export { default as ForgotPassword } from "./components/forgotPassword";
 export { default as GroupForm } from "./components/groupForm";
 export { default as GroupHeader } from "./components/groupHeader";
 export { default as GroupsTable } from "./components/groupsTable";
-export { default as GroupsTableCell } from "./components/groupsTableCell";
 export { default as GroupsTableButton } from "./components/groupsTableButton";
+export { default as GroupsTableCell } from "./components/groupsTableCell";
+export { default as Login } from "./components/login";
+export { default as LoginButtons } from "./components/loginButtons";
+export { default as LoginFormMessages } from "./components/loginFormMessages";
+export { default as MainDropdown } from "./components/mainDropdown";
 export { default as ManageGroups } from "./components/manageGroups";
 export { default as PermissionsList } from "./components/permissionsList";
+export { default as SignIn } from "./components/signIn";
+export { default as SignUp } from "./components/signUp";
+export { default as UpdatePasswordOverlay } from "./components/updatePasswordOverlay";
 
 export { default as AccountsDashboardContainer } from "./containers/accountsDashboardContainer";
+export { default as AuthContainer } from "./containers/auth";
 export { default as EditGroupContainer } from "./containers/editGroupContainer";
+export { default as ForgotPasswordContainer } from "./containers/forgotPassword";
+export { default as MainDropdownContainer } from "./containers/mainDropdown";
+export { default as MessagesContainer } from "./containers/messages";
+export { default as UpdatePasswordOverlayContainer } from "./containers/passwordOverlay";
+
+import "./templates/accounts.html";
+
+import "./templates/addressBook/add/add.html";
+import "./templates/addressBook/add/add.js";
+import "./templates/addressBook/edit/edit.html";
+import "./templates/addressBook/edit/edit.js";
+import "./templates/addressBook/form/form.html";
+import "./templates/addressBook/form/form.js";
+import "./templates/addressBook/grid/grid.html";
+import "./templates/addressBook/grid/grid.js";
+import "./templates/addressBook/review/review.html";
+import "./templates/addressBook/review/review.js";
+import "./templates/addressBook/addressBook.html";
+import "./templates/addressBook/addressBook.js";
+import "./templates/dashboard/dashboard.html";
+import "./templates/dashboard/dashboard.js";
+import "./templates/dropdown/helpers";
+import "./templates/inline/inline.html";
+import "./templates/inline/inline.js";
+import "./templates/login/loginForm.html";
+import "./templates/login/loginForm.js";
+import "./templates/members/member.html";
+import "./templates/members/member.js";
+import "./templates/profile/profile.html";
+import "./templates/profile/profile.js";
+import "./templates/updatePassword/updatePassword.html";
+import "./templates/updatePassword/updatePassword.js";
+import "./templates/verify/verifyAccount.html";
+import "./templates/verify/verifyAccount.js";
