+++ imports/plugins/core/accounts/client/templates/accounts.html
