+++ imports/plugins/core/accounts/client/templates/addressBook/addressBook.js
