+++ imports/plugins/core/accounts/client/templates/dashboard/dashboard.js
