+++ imports/plugins/core/accounts/client/templates/login/loginForm.js
