+++ imports/plugins/core/accounts/client/templates/updatePassword/updatePassword.html
