+++ imports/plugins/core/accounts/client/templates/verify/verifyAccount.html
